+++ src/integration-test/java/com/pyranid/MySqlIntegrationIT.java
@@ -330,6 +330,12 @@ public class MySqlIntegrationIT extends AbstractPortableJdbcIntegrationTests {
 			String autoIncrementPrimaryKey(@NonNull String columnName) {
 				return columnName + " BIGINT AUTO_INCREMENT PRIMARY KEY";
 			}
+
+			@NonNull
+			@Override
+			String timestampWithFractionalSeconds() {
+				return "DATETIME(6)";
+			}
 		};
 	}